+++ pytorch_lightning/trainer/trainer.py
@@ -1272,7 +1272,7 @@ class Trainer(
             profiler = profiler.lower()
             if profiler not in PROFILERS:
                 raise MisconfigurationException(
-                    "When passing string value for the `profiler` parameter of `Trainer, "
+                    "When passing string value for the `profiler` parameter of `Trainer`,"
                     f" it can only be one of {list(PROFILERS.keys())}"
                 )
             profiler_class = PROFILERS[profiler]
